refactor search dao and remove unused routes
