+++ wherehows-frontend/app/controllers/api/v1/Search.java
@@ -14,7 +14,6 @@
 package controllers.api.v1;
 
 import com.fasterxml.jackson.databind.node.ObjectNode;
-import dao.SearchDAO;
 import java.util.List;
 import org.apache.commons.lang3.math.NumberUtils;
 import play.Logger;
@@ -26,12 +25,12 @@ import play.mvc.Result;
 
 import static org.apache.commons.lang3.StringUtils.*;
 
+import wherehows.dao.table.SearchDAO;
+
 
 public class Search extends Controller {
   private static final String AUTOCOMPLETE_ALL_KEY = "autocomplete.all";
   private static final String AUTOCOMPLETE_DATASET_KEY = "autocomplete.dataset";
-  private static final String AUTOCOMPLETE_FLOW_KEY = "autocomplete.flow";
-  private static final String AUTOCOMPLETE_METRIC_KEY = "autocomplete.metric";
   private static final int DEFAULT_AUTOCOMPLETE_SIZE = 20;
   private static final int DEFAULT_AUTOCOMPLETE_CACHE_TIME = 3600; // cache for an hour
 
@@ -79,49 +78,7 @@ public class Search extends Controller {
     return ok(result);
   }
 
-  public static Result getSearchAutoCompleteForMetric() {
-    // if not input, then get all search names (without limit).
-    String input = request().getQueryString("input");
-    int size = 0;  // size 0 means no limit
-    if (isNotBlank(input)) {
-      size = NumberUtils.toInt(request().getQueryString("size"), DEFAULT_AUTOCOMPLETE_SIZE);
-    }
 
-    String cacheKey = AUTOCOMPLETE_METRIC_KEY + (isNotBlank(input) ? "." + input : "-all");
-    List<String> names = (List<String>) Cache.get(cacheKey);
-    if (names == null || names.size() == 0) {
-      names = SearchDAO.getAutoCompleteListMetric(input, size);
-      Cache.set(cacheKey, names, DEFAULT_AUTOCOMPLETE_CACHE_TIME);
-    }
-
-    ObjectNode result = Json.newObject();
-    result.put("status", "ok");
-    result.put("input", input);
-    result.set("source", Json.toJson(names));
-    return ok(result);
-  }
-
-  public static Result getSearchAutoCompleteForFlow() {
-    // if not input, then get all search names (without limit).
-    String input = request().getQueryString("input");
-    int size = 0;  // size 0 means no limit
-    if (isNotBlank(input)) {
-      size = NumberUtils.toInt(request().getQueryString("size"), DEFAULT_AUTOCOMPLETE_SIZE);
-    }
-
-    String cacheKey = AUTOCOMPLETE_FLOW_KEY + (isNotBlank(input) ? "." + input : "-all");
-    List<String> names = (List<String>) Cache.get(cacheKey);
-    if (names == null || names.size() == 0) {
-      names = SearchDAO.getAutoCompleteListFlow(input, size);
-      Cache.set(cacheKey, names, DEFAULT_AUTOCOMPLETE_CACHE_TIME);
-    }
-
-    ObjectNode result = Json.newObject();
-    result.put("status", "ok");
-    result.put("input", input);
-    result.set("source", Json.toJson(names));
-    return ok(result);
-  }
 
   public static Result searchByKeyword() {
     ObjectNode result = Json.newObject();
@@ -165,38 +122,9 @@ public class Search extends Controller {
     }
 
     String searchEngine = Play.application().configuration().getString(SearchDAO.WHEREHOWS_SEARCH_ENGINE_KEY);
+    Logger.info("searchEngine is: " + searchEngine); // TODO: deprecated this setting
 
-    if (category.toLowerCase().equalsIgnoreCase("metrics")) {
-      if (isNotBlank(searchEngine) && searchEngine.equalsIgnoreCase("elasticsearch")) {
-        result.set("result", SearchDAO.elasticSearchMetricByKeyword(category, keyword, page, size));
-      } else {
-        result.set("result", SearchDAO.getPagedMetricByKeyword(category, keyword, page, size));
-      }
-    } else if (category.toLowerCase().equalsIgnoreCase("flows")) {
-      if (isNotBlank(searchEngine) && searchEngine.equalsIgnoreCase("elasticsearch")) {
-        result.set("result", SearchDAO.elasticSearchFlowByKeyword(category, keyword, page, size));
-      } else {
-        result.set("result", SearchDAO.getPagedFlowByKeyword(category, keyword, page, size));
-      }
-    } else if (category.toLowerCase().equalsIgnoreCase("jobs")) {
-      if (isNotBlank(searchEngine) && searchEngine.equalsIgnoreCase("elasticsearch")) {
-        result.set("result", SearchDAO.elasticSearchFlowByKeyword(category, keyword, page, size));
-      } else {
-        result.set("result", SearchDAO.getPagedJobByKeyword(category, keyword, page, size));
-      }
-    } else if (category.toLowerCase().equalsIgnoreCase("comments")) {
-      if (isNotBlank(searchEngine) && searchEngine.equalsIgnoreCase("elasticsearch")) {
-        result.set("result", SearchDAO.elasticSearchDatasetByKeyword(category, keyword, null, page, size));
-      } else {
-        result.set("result", SearchDAO.getPagedCommentsByKeyword(category, keyword, page, size));
-      }
-    } else {
-      if (isNotBlank(searchEngine) && searchEngine.equalsIgnoreCase("elasticsearch")) {
-        result.set("result", SearchDAO.elasticSearchDatasetByKeyword(category, keyword, source, page, size));
-      } else {
-        result.set("result", SearchDAO.getPagedDatasetByKeyword(category, keyword, source, page, size));
-      }
-    }
+    result.set("result", SearchDAO.elasticSearchDatasetByKeyword(category, keyword, source, page, size));
 
     return ok(result);
   }
